Ticket RHEVM-808:
java sdk backend: wrong conversion of BigDecimal

Reviewed-on: https://gerrit.eng.lab.tlv.redhat.com/10620
Tested-by: Ilia Meerovich <imeerovi@redhat.com>

+++ art/core_api/ovirtsdk_java_utils.py
@@ -23,6 +23,7 @@ import types
 from functools import wraps
 import threading
 import thread
+from decimal import Decimal
 
 from art.core_api import validator
 from art.core_api.apis_exceptions import EntityNotFound
@@ -752,7 +753,7 @@ class JavaTranslator(object):
             python_data = 'true' if data is True else 'false'
         # BigDecimal case
         elif isinstance(data, java.math.BigDecimal):
-            python_data = int(data.toString())
+            python_data = Decimal(data.toString())
         # assuming that this is string
         elif data is not None:
             python_data = data.encode('utf8')
